+++ src/Main.cc
@@ -2,15 +2,15 @@
 #include <iostream>
 
 // All Components
-#include "CollisionComp.h"
-#include "InputComp.h"
-#include "LifespanComp.h"
-#include "ScoreComp.h"
-#include "ShapeComp.h"
-#include "TransformComp.h"
+#include "EcsHeaders/CollisionComp.h"
+#include "EcsHeaders/InputComp.h"
+#include "EcsHeaders/LifespanComp.h"
+#include "EcsHeaders/ScoreComp.h"
+#include "EcsHeaders/ShapeComp.h"
+#include "EcsHeaders/TransformComp.h"
 
 // Entity classes
-#include "EntityManager.h"
+#include "EcsHeaders/EntityManager.h"
 
 // For ease of use, use both std & our custom ecse namespaces
 using namespace std;
